+++ e2e/pom/planning-poker/JoinSessionPagePom.ts
